Remove `UserController` and related classes, interfaces, and tests
